deploy: b90bcf35b14852f989384e715757a02a2c35e72b

+++ examples/event_queue/event_queue.c
@@ -76,7 +76,7 @@ int main(void) {
                     printf("Drag : %i %i\n", event.drag.x, event.drag.y);
                     break;
                 case RGFW_scaleUpdated:
-                    printf("Scale Updated : %f %f\n", event.scale.x, event.scale.y);
+                    printf("Scale Updated : %f %f\n", (double)event.scale.x, (double)event.scale.y);
                     break;
                 default:
                     break;
